Updated: Components that served as features has been refactored to the features folder
