+++ src/Pages/Home/landingPageProduct/Beauty.jsx
