+++ src/Pages/productCategories/SingleProductPage.jsx
@@ -3,7 +3,7 @@ import { ChevronRight } from "lucide-react";
 import { useNavigate, useParams } from "react-router";
 import { ProductCategoriesShortcut } from "./ProductCategoriesShortcut";
 import { SingleProductDetails } from "./productDetails/SingleProductDetails";
-import { MiniProductCategories } from "../../ui/components/hero/MiniProductCategories";
+import { MiniProductCategories } from "../home/hero/MiniProductCategories";
 
 const SingleProductPage = () => {
   const { categoryName, subcategory, slug, id } = useParams();
