+++ src/features/cart/cartSlice.js
@@ -56,10 +56,10 @@ export default cartSlice.reducer;
 export const getCart = (state) => state.cart.cart;
 
 export const getTotalCartQuantity = (state) =>
-  state.cart.cart.reduce((sum, item) => sum + item.quantity, 0);
+  state.cart.cart.reduce((sum, item) => sum + item?.quantity, 0);
 
 export const getTotalCartPrice = (state) =>
-  state.cart.cart.reduce((sum, item) => sum + item.totalPrice, 0);
+  state.cart.cart.reduce((sum, item) => sum + item?.totalPrice, 0);
 
 export const getCurrentQuantityById = (id) => (state) =>
   state.cart.cart.find((item) => item.id === id)?.quantity ?? 0;
