+++ src/features/reviewsRating/ratingSlice.js
@@ -1,0 +1,3 @@
+const InitialState = {
+  rating: 5,
+};
